Update tests to use latest DOM release
