+++ test/unit/HTMLDocumentTest.php
@@ -4,7 +4,6 @@ namespace Gt\DomTemplate\Test;
 use Gt\DomTemplate\DocumentFragment;
 use Gt\DomTemplate\Element;
 use Gt\DomTemplate\HTMLDocument;
-use Gt\DomTemplate\Test\Helper\Helper;
 
 class HTMLDocumentTest extends TestCase {
 	public function testOverriddenClasses() {
@@ -18,12 +17,5 @@ class HTMLDocumentTest extends TestCase {
 		$document = new HTMLDocument("<!doctype html><h1 id='test'>Test</h1>");
 		$element = $document->getElementById("test");
 		self::assertInstanceOf(Element::class, $element);
-	}
-
-	public function testExtractTemplatesSetsParentInnerHTMLToEmpty() {
-		$document = new HTMLDocument(Helper::HTML_TODO_LIST);
-		$document->extractTemplates();
-		$todoListElement = $document->getElementById("todo-list");
-		self::assertSame("", $todoListElement->innerHTML);
 	}
 }
